Step 13: Leverage 3 layers, DAL implementation (repositories and execute Unit of Work patterns)

+++ QuizApp.Data/Repositories/QuizRepository.cs
@@ -28,6 +28,11 @@ public class QuizRepository : IQuizRepository
 
     public bool Delete(Quiz entity)
     {
+        if (entity == null)
+        {
+            return false; // Không có gì để xóa
+        }
+
         _context.Quizzes.Remove(entity);
         return _context.SaveChanges() > 0;
     }
